minor bugfix
After AUR package installation step, lemonaid parses lemonaid.log for strings containing 'AUR' to determine if an aur package failed and report that errors have been detected if so. The specific wording of a log write if RETRYASAUR is enabled would cause errors to be falsely reported during the AUR package installation step when in fact none had occured.